refactor: findWinnder 함수 findingWinners.js 파일로 이동
