--- src/Models/winnerChoice.js
@@ -1,2 +1,0 @@
-
-
