+++ src/Views/outputView.js
@@ -5,11 +5,11 @@ const printNotificationExecutionResult = () => {
 };
 
 const printRacingProgressOutput = (carNamesInput, progressionLengthForEachRacer) => {
-    for (let i = 0; i < carNamesInput.length; i += 1) {
-      Console.print(`${carNamesInput[i]} : ${progressionLengthForEachRacer[i]}`);
-    }
-    Console.print('\n');
-  };
+  for (let i = 0; i < carNamesInput.length; i += 1) {
+    Console.print(`${carNamesInput[i]} : ${progressionLengthForEachRacer[i]}`);
+  }
+  Console.print('\n');
+};
 
 const printFinalWinners = MaxNumberNames => {
   Console.print('최종 우승자 : ' + MaxNumberNames.join(', '));
